Removed MX record check in email validate function.
According to RFC5321 Section 5.1:
If an empty list of MXs is returned,
   the address is treated as if it was associated with an implicit MX
   RR, with a preference of 0, pointing to that host.  If MX records are
   present, but none of them are usable, or the implicit MX is unusable,
   this situation MUST be reported as an error.

Thus we cannot use checkdnsrr($dom, 'MX'), because it should only fail when there are multiple MX records which this function doesn't support.

We can check the count of the MX records with dns_get_record($dom, DNS_MX). The count should be 1 or 0.

An addition could be checking if the MX records found actually exist. But this is probably overkill.